requirements.txt
aiohttp==3.4.4
amqp==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
flower==0.9.2
gidgethub==3.0.0
github3.py==1.2.0
idna==2.7
idna-ssl==1.1.0
jwcrypto==0.5.0
kombu==4.2.1
multidict==4.4.2
pycparser==2.19
python-dateutil==2.7.3
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
six==1.11.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
yarl==1.2.6
